EOS-24293: Created a separate api to delete queued task
